refactored views
